refactor: replace blockExplorer hook by plain TS functions
