+++ packages/lib/modules/pool/PoolDetail/PoolActivityChart/usePoolActivityChart.tsx
@@ -11,10 +11,6 @@ import { ChainSlug, getChainSlug } from '../../pool.utils'
 import { ColorMode, useTheme as useChakraTheme } from '@chakra-ui/react'
 import { useTheme as useNextTheme } from 'next-themes'
 import { abbreviateAddress } from '@repo/lib/shared/utils/addresses'
-import {
-  getBlockExplorerAddressUrl,
-  getBlockExplorerTxUrl,
-} from '@repo/lib/shared/hooks/useBlockExplorer'
 import { useBreakpoints } from '@repo/lib/shared/hooks/useBreakpoints'
 import { useCurrency } from '@repo/lib/shared/hooks/useCurrency'
 import { NumberFormatter } from '@repo/lib/shared/utils/numbers'
@@ -24,6 +20,10 @@ import {
   PoolActivityTokens,
   PoolActivityEl,
 } from '../PoolActivity/poolActivity.types'
+import {
+  getBlockExplorerAddressUrl,
+  getBlockExplorerTxUrl,
+} from '@repo/lib/shared/utils/blockExplorer'
 
 const getDefaultPoolActivityChartOptions = (
   nextTheme: ColorMode = 'dark',
