+++ packages/lib/modules/pool/PoolDetail/PoolUserEvents/PoolUserEvents.tsx
@@ -23,13 +23,13 @@ import {
 } from '@repo/lib/shared/services/api/generated/graphql'
 import { TokenIcon } from '@repo/lib/modules/tokens/TokenIcon'
 import { formatDistanceToNow, secondsToMilliseconds } from 'date-fns'
-import { useBlockExplorer } from '@repo/lib/shared/hooks/useBlockExplorer'
 import { ArrowUpRight } from 'react-feather'
 import { PoolEventItem } from '../../usePoolEvents'
 import { calcTotalStakedBalance, getUserTotalBalance } from '../../user-balance.helpers'
 import { fNum, bn } from '@repo/lib/shared/utils/numbers'
 import { isEmpty } from 'lodash'
 import { BoostText } from './BoostText'
+import { getBlockExplorerTxUrl } from '@repo/lib/shared/utils/blockExplorer'
 
 type PoolEventRowProps = {
   poolEvent: PoolEventItem
@@ -183,7 +183,6 @@ export default function PoolUserEvents({
   const [height, setHeight] = useState(0)
   const [poolEvents, setPoolEvents] = useState<PoolEventItem[]>([])
   const { toCurrency } = useCurrency()
-  const { getBlockExplorerTxUrl } = useBlockExplorer(chain)
 
   const isVeBal = pool.staking?.type === GqlPoolStakingType.Vebal
   const showBoostValue = pool.staking?.type === GqlPoolStakingType.Gauge && !isVeBal
