+++ packages/lib/shared/components/BlockExplorerLink.tsx
@@ -2,9 +2,9 @@
 
 import { HStack, Link, Text } from '@chakra-ui/react'
 import { ArrowUpRight } from 'react-feather'
-import { getBlockExplorerTxUrl, getBlockExplorerName } from '../hooks/useBlockExplorer'
 import { Address } from 'viem'
 import { GqlChain } from '../services/api/generated/graphql'
+import { getBlockExplorerName, getBlockExplorerTxUrl } from '../utils/blockExplorer'
 
 type Props = { transactionHash?: Address; chain: GqlChain }
